factor 4 (#531)
(cherry picked from commit 4cea6218ac03505aa705da61a8970a19e505568e)
