+++ python/sglang/jit_kernel/dsv4/online_c128_mtp.py
@@ -1,60 +1,14 @@
 from __future__ import annotations
 
+from dataclasses import dataclass
+from typing import Any, List, Optional, Tuple
+
 import torch
-import triton
-import triton.language as tl
 from tvm_ffi.module import Module
 
 from sglang.jit_kernel.dsv4.utils import make_name
 from sglang.jit_kernel.utils import cache_once, load_jit, make_cpp_args
-
-
-@triton.jit
-def _online_c128_mtp_prepare_kernel(
-    seq_lens_ptr,
-    req_pool_indices_ptr,
-    req_to_token_ptr,
-    full_to_swa_ptr,
-    main_state_ptr,
-    req_to_token_stride_b: tl.constexpr,
-    main_state_stride_b: tl.constexpr,
-    bs,
-    swa_page_size: tl.constexpr,
-    temp_state_slot_offset: tl.constexpr,
-    state_width: tl.constexpr,
-    block_d: tl.constexpr,
-):
-    bid = tl.program_id(0)
-    if bid >= bs:
-        return
-
-    d = tl.arange(0, block_d)
-    d_mask = d < state_width
-    seq_len = tl.load(seq_lens_ptr + bid).to(tl.int64)
-    has_partial = (seq_len > 0) & ((seq_len % 128) != 0)
-
-    chunk_start = tl.where(has_partial, ((seq_len - 1) // 128) * 128, 0)
-    req_idx = tl.load(req_pool_indices_ptr + bid).to(tl.int64)
-    full_loc = tl.load(
-        req_to_token_ptr + req_idx * req_to_token_stride_b + chunk_start,
-        mask=has_partial,
-        other=0,
-    )
-    swa_loc = tl.load(full_to_swa_ptr + full_loc, mask=has_partial, other=0).to(
-        tl.int64
-    )
-    slot = tl.where(has_partial, swa_loc // swa_page_size, 0)
-    temp_slot = slot + temp_state_slot_offset
-    value = tl.load(
-        main_state_ptr + slot * main_state_stride_b + d,
-        mask=d_mask & has_partial,
-        other=0.0,
-    )
-    tl.store(
-        main_state_ptr + temp_slot * main_state_stride_b + d,
-        value,
-        mask=d_mask & has_partial,
-    )
+from sglang.srt.environ import envs
 
 
 @cache_once
@@ -66,7 +20,9 @@ def _jit_online_c128_mtp_module(head_dim: int) -> Module:
         cuda_files=["deepseek_v4/online_c128_mtp.cuh"],
         cuda_wrappers=[
             ("write_prefix_states", f"OnlineC128MTPWritePrefixKernel<{args}>::run"),
-            ("lazy_commit", f"OnlineC128MTPLazyCommitKernel<{args}>::run"),
+            ("mark_pending", f"OnlineC128MTPMarkPendingKernel<{args}>::run"),
+            ("commit_pending", f"OnlineC128MTPCommitPendingKernel<{args}>::run"),
+            ("clear_pending", f"OnlineC128MTPClearPendingKernel<{args}>::run"),
         ],
         extra_cuda_cflags=["-use_fast_math"],
     )
@@ -104,92 +60,314 @@ def online_c128_mtp_write_prefix_states(
     )
 
 
-def online_c128_mtp_lazy_commit(
+def online_c128_mtp_mark_pending(
     *,
-    old_seq_lens: torch.Tensor,
-    old_req_pool_indices: torch.Tensor,
+    seq_lens: torch.Tensor,
+    req_pool_indices: torch.Tensor,
+    pending_seq_lens: torch.Tensor,
+    bs: int,
+    max_num_reqs: int,
+    head_dim: int,
+) -> None:
+    if bs <= 0:
+        return
+    _jit_online_c128_mtp_module(head_dim).mark_pending(
+        seq_lens,
+        req_pool_indices,
+        pending_seq_lens,
+        bs,
+        max_num_reqs,
+    )
+
+
+def online_c128_mtp_commit_pending(
+    *,
     cur_seq_lens: torch.Tensor,
     cur_req_pool_indices: torch.Tensor,
     req_to_token: torch.Tensor,
     full_to_swa_index_mapping: torch.Tensor,
+    pending_seq_lens: torch.Tensor,
     state: torch.Tensor,
-    old_bs: int,
     cur_bs: int,
     swa_page_size: int,
     num_verify_tokens: int,
     state_slot_stride: int,
+    max_num_reqs: int,
     head_dim: int,
 ) -> None:
-    if old_bs <= 0 or cur_bs <= 0:
+    if cur_bs <= 0:
         return
-    _jit_online_c128_mtp_module(head_dim).lazy_commit(
-        old_seq_lens,
-        old_req_pool_indices,
+    _jit_online_c128_mtp_module(head_dim).commit_pending(
         cur_seq_lens,
         cur_req_pool_indices,
         req_to_token,
         full_to_swa_index_mapping,
+        pending_seq_lens,
         state,
-        old_bs,
         cur_bs,
         swa_page_size,
         num_verify_tokens,
         state_slot_stride,
+        max_num_reqs,
     )
 
 
-def _online_c128_mtp_prepare_triton(
+def online_c128_mtp_clear_pending(
     *,
-    seq_lens: torch.Tensor,
     req_pool_indices: torch.Tensor,
-    req_to_token: torch.Tensor,
-    full_to_swa_index_mapping: torch.Tensor,
-    main_state: torch.Tensor,
+    pending_seq_lens: torch.Tensor,
     bs: int,
-    swa_page_size: int,
-    temp_state_slot_offset: int,
-    state_width: int,
+    head_dim: int,
 ) -> None:
     if bs <= 0:
         return
-    assert state_width <= 2048
-    block_d = triton.next_power_of_2(state_width)
-    _online_c128_mtp_prepare_kernel[(bs,)](
-        seq_lens,
+    _jit_online_c128_mtp_module(head_dim).clear_pending(
         req_pool_indices,
-        req_to_token,
-        full_to_swa_index_mapping,
-        main_state,
-        req_to_token.stride(0),
-        main_state.stride(0),
+        pending_seq_lens,
         bs,
-        swa_page_size,
-        temp_state_slot_offset,
-        state_width,
-        block_d,
     )
 
 
-def online_c128_mtp_prepare(
-    *,
-    seq_lens: torch.Tensor,
-    req_pool_indices: torch.Tensor,
-    req_to_token: torch.Tensor,
-    full_to_swa_index_mapping: torch.Tensor,
-    main_state: torch.Tensor,
-    bs: int,
-    swa_page_size: int,
-    temp_state_slot_offset: int,
-    state_width: int,
-) -> None:
-    _online_c128_mtp_prepare_triton(
-        seq_lens=seq_lens,
-        req_pool_indices=req_pool_indices,
-        req_to_token=req_to_token,
-        full_to_swa_index_mapping=full_to_swa_index_mapping,
-        main_state=main_state,
-        bs=bs,
-        swa_page_size=swa_page_size,
-        temp_state_slot_offset=temp_state_slot_offset,
-        state_width=state_width,
-    )
+@dataclass
+class _OnlineC128LayerRuntime:
+    head_dim: int
+    main_state: torch.Tensor
+    state_slot_offset: int
+
+
+@dataclass
+class _OnlineC128VerifyContext:
+    req_pool_indices: torch.Tensor
+    seq_lens: torch.Tensor
+
+
+class OnlineC128MTPController:
+    def __init__(self, backend: Any):
+        self.backend = backend
+        self._verify_ctx: Optional[_OnlineC128VerifyContext] = None
+        self._compressors: Optional[List[Tuple[int, Any]]] = None
+        self._layer_runtimes: Optional[List[_OnlineC128LayerRuntime]] = None
+
+    def enabled(self) -> bool:
+        return (
+            envs.SGLANG_OPT_USE_ONLINE_COMPRESS.get()
+            and envs.SGLANG_EXPERIMENTAL_ONLINE_C128_MTP.get()
+            and self.backend.mtp_enabled
+        )
+
+    def state_slot_offset(self) -> int:
+        if not self.enabled():
+            return 0
+        return self.backend.token_to_kv_pool.get_online_c128_mtp_state_slot_offset()
+
+    def begin_verify(
+        self,
+        req_pool_indices: torch.Tensor,
+        seq_lens: torch.Tensor,
+    ) -> None:
+        if not self.enabled():
+            self.clear()
+            return
+
+        self._verify_ctx = _OnlineC128VerifyContext(
+            req_pool_indices=req_pool_indices.detach(),
+            seq_lens=seq_lens.detach(),
+        )
+        head_dim = self._head_dim()
+        if head_dim is None or self._num_verify_tokens() == 0:
+            return
+        token_to_kv_pool = self.backend.token_to_kv_pool
+        online_c128_mtp_mark_pending(
+            seq_lens=seq_lens,
+            req_pool_indices=req_pool_indices,
+            pending_seq_lens=token_to_kv_pool.get_online_c128_mtp_pending_seq_lens(),
+            bs=min(seq_lens.shape[0], req_pool_indices.shape[0]),
+            max_num_reqs=token_to_kv_pool.max_num_reqs,
+            head_dim=head_dim,
+        )
+
+    def clear(self) -> None:
+        self._verify_ctx = None
+
+    def prepare_forward(
+        self,
+        logical_forward_mode,
+        req_pool_indices: torch.Tensor,
+        seq_lens: torch.Tensor,
+        *,
+        verify_bs: Optional[int] = None,
+    ) -> int:
+        if not self.enabled():
+            self.clear()
+            return 0
+        if logical_forward_mode.is_idle():
+            self.clear()
+            return 0
+
+        active_req_pool_indices = req_pool_indices
+        active_seq_lens = seq_lens
+        if logical_forward_mode.is_target_verify():
+            if verify_bs is None:
+                verify_bs = req_pool_indices.shape[0]
+            active_req_pool_indices = req_pool_indices[:verify_bs]
+            active_seq_lens = seq_lens[:verify_bs]
+            if verify_bs == 0:
+                self.clear()
+                return 0
+
+        self.commit_pending(
+            req_pool_indices=active_req_pool_indices,
+            seq_lens=active_seq_lens,
+        )
+        if not logical_forward_mode.is_target_verify():
+            return 0
+
+        self.begin_verify(
+            req_pool_indices=active_req_pool_indices,
+            seq_lens=active_seq_lens,
+        )
+        return self.state_slot_offset()
+
+    def write_prefix_states(
+        self,
+        layer_id: int,
+        compressor: Any,
+        kv_score_input: torch.Tensor,
+        logical_forward_mode,
+    ) -> None:
+        if (
+            not self.enabled()
+            or not logical_forward_mode.is_target_verify()
+            or compressor.is_in_indexer
+            or compressor.ratio != 128
+            or kv_score_input.numel() == 0
+        ):
+            return
+
+        ctx = self._active_ctx()
+        num_verify_tokens = self._num_verify_tokens()
+        if ctx is None or num_verify_tokens == 0:
+            return
+
+        token_to_kv_pool = self.backend.token_to_kv_pool
+        head_dim = compressor.head_dim
+        state_pool = token_to_kv_pool.get_attention_compress_states(layer_id)
+        total_bs = kv_score_input.numel() // (num_verify_tokens * head_dim * 2)
+        layer_bs = min(ctx.seq_lens.shape[0], ctx.req_pool_indices.shape[0], total_bs)
+        if layer_bs <= 0:
+            return
+
+        online_c128_mtp_write_prefix_states(
+            kv_score_input=kv_score_input,
+            seq_lens=ctx.seq_lens,
+            req_pool_indices=ctx.req_pool_indices,
+            req_to_token=self.backend.req_to_token,
+            full_to_swa_index_mapping=token_to_kv_pool.full_to_swa_index_mapping,
+            ape=compressor.ape.reshape(128, head_dim),
+            state=state_pool.kv_score_buffer.kv_score,
+            layer_bs=layer_bs,
+            swa_page_size=token_to_kv_pool.swa_page_size,
+            num_verify_tokens=num_verify_tokens,
+            state_slot_stride=state_pool.online_mtp_state_slot_offset,
+            head_dim=head_dim,
+        )
+
+    def commit_pending(
+        self,
+        req_pool_indices: torch.Tensor,
+        seq_lens: torch.Tensor,
+    ) -> None:
+        if self._verify_ctx is None:
+            return
+        if not self.enabled():
+            self.clear()
+            return
+        if req_pool_indices.numel() == 0 or seq_lens.numel() == 0:
+            return
+
+        num_verify_tokens = self._num_verify_tokens()
+        if num_verify_tokens == 0:
+            self.clear()
+            return
+
+        backend = self.backend
+        token_to_kv_pool = backend.token_to_kv_pool
+        pending_seq_lens = token_to_kv_pool.get_online_c128_mtp_pending_seq_lens()
+        cur_bs = min(seq_lens.shape[0], req_pool_indices.shape[0])
+
+        for runtime in self._iter_layer_runtimes():
+            online_c128_mtp_commit_pending(
+                cur_seq_lens=seq_lens,
+                cur_req_pool_indices=req_pool_indices,
+                req_to_token=backend.req_to_token,
+                full_to_swa_index_mapping=token_to_kv_pool.full_to_swa_index_mapping,
+                pending_seq_lens=pending_seq_lens,
+                state=runtime.main_state,
+                cur_bs=cur_bs,
+                swa_page_size=token_to_kv_pool.swa_page_size,
+                num_verify_tokens=num_verify_tokens,
+                state_slot_stride=runtime.state_slot_offset,
+                max_num_reqs=token_to_kv_pool.max_num_reqs,
+                head_dim=runtime.head_dim,
+            )
+
+        head_dim = self._head_dim()
+        if head_dim is not None:
+            online_c128_mtp_clear_pending(
+                req_pool_indices=req_pool_indices,
+                pending_seq_lens=pending_seq_lens,
+                bs=cur_bs,
+                head_dim=head_dim,
+            )
+        self.clear()
+
+    def _num_verify_tokens(self) -> int:
+        if not self.enabled():
+            return 0
+        num_verify_tokens = int(self.backend.speculative_num_draft_tokens)
+        max_draft_tokens = (
+            self.backend.token_to_kv_pool.get_online_c128_mtp_max_draft_tokens()
+        )
+        return num_verify_tokens if 0 < num_verify_tokens <= max_draft_tokens else 0
+
+    def _active_ctx(self) -> Optional[_OnlineC128VerifyContext]:
+        ctx = self._verify_ctx
+        if (
+            ctx is None
+            or ctx.seq_lens.numel() == 0
+            or ctx.req_pool_indices.numel() == 0
+        ):
+            return None
+        return ctx
+
+    def _head_dim(self) -> Optional[int]:
+        for runtime in self._iter_layer_runtimes():
+            return runtime.head_dim
+        return None
+
+    def _iter_online_c128_layers(self):
+        if self._compressors is None:
+            compressors = []
+            for layer in self.backend.model_runner.model.model.layers:
+                attn = getattr(layer, "self_attn", None)
+                compressor = getattr(attn, "compressor", None)
+                if compressor is not None and compressor.ratio == 128:
+                    compressors.append((compressor.layer_id, compressor))
+            self._compressors = compressors
+        return iter(self._compressors)
+
+    def _iter_layer_runtimes(self):
+        if self._layer_runtimes is None:
+            runtimes = []
+            token_to_kv_pool = self.backend.token_to_kv_pool
+            for layer_id, compressor in self._iter_online_c128_layers():
+                state_pool = token_to_kv_pool.get_attention_compress_states(layer_id)
+                runtimes.append(
+                    _OnlineC128LayerRuntime(
+                        head_dim=compressor.head_dim,
+                        main_state=state_pool.kv_score_buffer.kv_score,
+                        state_slot_offset=state_pool.online_mtp_state_slot_offset,
+                    )
+                )
+            self._layer_runtimes = runtimes
+        return iter(self._layer_runtimes)
