+++ python/sglang/srt/mem_cache/deepseek_v4_memory_pool.py
@@ -406,6 +406,11 @@ class DeepSeekV4TokenToKVPool(BaseSWAKVPool):
         self.c128_state_pool_size = c128_state_pool_size
         self.state_dtype = state_dtype
         self.compression_ratios = compression_ratios
+        self.online_c128_mtp_pending_seq_lens: Optional[torch.Tensor] = None
+        if ONLINE_C128 and envs.SGLANG_EXPERIMENTAL_ONLINE_C128_MTP.get():
+            self.online_c128_mtp_pending_seq_lens = torch.empty(
+                max_num_reqs, dtype=torch.int64, device=device
+            )
 
         # Determine this PP stage's absolute layer range
         if (
@@ -653,6 +658,10 @@ class DeepSeekV4TokenToKVPool(BaseSWAKVPool):
             if pool is not None and pool.ratio == 128:
                 return int(pool.online_mtp_max_draft_tokens)
         return 0
+
+    def get_online_c128_mtp_pending_seq_lens(self) -> torch.Tensor:
+        assert self.online_c128_mtp_pending_seq_lens is not None
+        return self.online_c128_mtp_pending_seq_lens
 
     def get_indexer_compress_states(self, layer_id: int) -> CompressStatePool:
         self.wait_layer_transfer(layer_id)
